Modules: support `CMAKE_Swift_COMPILER_TARGET`
This allows the user to specify the target of the compiled output by the
standard CMake variable `CMAKE_Swift_COMPILER_TARGET`.  This requires
refactoring the emission of the frontend flag and a couple of
workarounds for the linker invocation.  Because `CMAKE_Swift_COMPILER`
will add the `CMAKE_Swift_COMPILER_ARG1` implicitly, we need to use the
`CMAKE_Swift_COMPILER` variable instead of the placeholder.  Then, we
need to add the `-target` value ourselves to ensure that the linker is
invoked properly by the linker driver (i.e. `clang`).  This improves the
ability to cross-compile Swift code with CMake.